backend/requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
firebase-admin>=6.5.0
opencv-python>=4.9.0
mediapipe>=0.10.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
Pillow>=10.3.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
